# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tienda Magic API",
        "description": "Documentación de la API de Tienda Magic Swagger OpenApi",
        "contact": {
            "email": "soporte@tiendamagic.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/decks": {
            "get": {
                "tags": [
                    "Decks"
                ],
                "summary": "Mis mazos",
                "description": "Lista los mazos del usuario autenticado.",
                "operationId": "d05f15850eef79716d6ed2e4322131ab",
                "responses": {
                    "200": {
                        "description": "Mazos listados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Decks"
                ],
                "summary": "Crear mazo",
                "description": "Crea un mazo nuevo.",
                "operationId": "f46b2d6188181c424a9b1d0b4fcd3c34",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "is_public": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mazo creado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/decks/{id}": {
            "get": {
                "tags": [
                    "Decks"
                ],
                "summary": "Ver mazo",
                "description": "Obtiene los detalles de un mazo y las cartas contenidas.",
                "operationId": "6ba360ec241a29ca842800aca0fbd262",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del mazo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mazo cargado exitosamente"
                    },
                    "403": {
                        "description": "Mazo privado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/decks/{deckId}/cards": {
            "post": {
                "tags": [
                    "Decks"
                ],
                "summary": "Añadir carta al mazo",
                "description": "Añade una cantidad de una carta a un mazo específico.",
                "operationId": "000bf7258b8396d5486390861763d1d7",
                "parameters": [
                    {
                        "name": "deckId",
                        "in": "path",
                        "description": "ID del mazo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "card_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carta añadida exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/decks/{deckId}/cards/{cardId}": {
            "delete": {
                "tags": [
                    "Decks"
                ],
                "summary": "Quitar carta del mazo",
                "description": "Elimina una carta de un mazo específico.",
                "operationId": "48d740fc7f25d9500d30cc488cdc7853",
                "parameters": [
                    {
                        "name": "deckId",
                        "in": "path",
                        "description": "ID del mazo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cardId",
                        "in": "path",
                        "description": "ID de la carta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carta eliminada exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/search/all": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Búsqueda global",
                "description": "Búsqueda unificada en toda la tienda y foro (stub).",
                "operationId": "df647696551cecae60b10acb71f8c2a1",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de búsqueda"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}